moved parts of mock to mock-related files
