+++ src/lib/temple/beacon.test.ts
@@ -1,6 +1,7 @@
 import { browser } from 'webextension-polyfill-ts';
 
-import './libsodium-wrappers.mock';
+// eslint-disable-next-line import/order
+import { mockSodiumUtil, MOCK_PK_KEY, MOCK_SK_KEY } from './libsodium-wrappers.mock';
 
 import {
   createCryptoBox,
@@ -30,7 +31,7 @@ import {
   encodeMessage,
   MessageType
 } from './beacon';
-import { mockBrowserStorageLocal, mockCryptoUtil, mockSodiumUtil, MOCK_PK_KEY, MOCK_SK_KEY } from './beacon.mock';
+import { mockBrowserStorageLocal, mockCryptoUtil } from './beacon.mock';
 
 browser.storage.local = { ...browser.storage.local, ...mockBrowserStorageLocal };
 global.crypto = { ...crypto, ...mockCryptoUtil };
